# docs/swagger.yaml
basePath: /
definitions:
  api.WalletRestoreRequest:
    properties:
      account_id:
        type: integer
      address_id:
        type: integer
      mnemonic:
        type: string
      password:
        type: string
      payment_id:
        type: integer
      stake_id:
        type: integer
    required:
    - mnemonic
    type: object
  bursa.KeyFile:
    properties:
      cborHex:
        type: string
      description:
        type: string
      type:
        type: string
    type: object
  bursa.Wallet:
    properties:
      mnemonic:
        type: string
      payment_address:
        type: string
      payment_extended_skey:
        $ref: '#/definitions/bursa.KeyFile'
      payment_skey:
        $ref: '#/definitions/bursa.KeyFile'
      payment_vkey:
        $ref: '#/definitions/bursa.KeyFile'
      stake_address:
        type: string
      stake_extended_skey:
        $ref: '#/definitions/bursa.KeyFile'
      stake_skey:
        $ref: '#/definitions/bursa.KeyFile'
      stake_vkey:
        $ref: '#/definitions/bursa.KeyFile'
    type: object
info:
  contact:
    email: support@blinklabs.io
    name: Blink Labs
    url: https://blinklabs.io
  description: Programmable Cardano Wallet API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: bursa
  version: v0
paths:
  /api/wallet/create:
    get:
      description: Create a wallet and return details
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/bursa.Wallet'
      summary: Create a wallet
  /api/wallet/restore:
    post:
      consumes:
      - application/json
      description: Restores a wallet using the provided mnemonic seed phrase and optional
        password and returns wallet details.
      parameters:
      - description: Wallet Restore Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.WalletRestoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Wallet successfully restored
          schema:
            $ref: '#/definitions/bursa.Wallet'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Restore a wallet using a mnemonic seed phrase
schemes:
- http
swagger: "2.0"